+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All important changes to `fiisoft-entity-indexer` will be documented in this file
 
+## 0.2.0
+
+* added class EntityMapStatsViewAggregator 
+* added new param to constructor of EntityMapWithStats
+* added method toArray() to EntityMapStatsView
+* added abstract class BaseEntityMapStatsView
+
 ## 0.1.0
 
 First version - need tests in real world...